refactor(web): scope /settings down to ESP32 controllers and set-password
W-21 (#4440). `/settings` keeps only what the web can do that the app
can't: the ESP32 controller list (Web-serial hardware that never moved to
the SPA) and set-password (mobile has reset-password, not set-password,
so an OAuth-only user has nowhere else to go). Everything else — profile,
grade format, logbook shortcut, board accounts, watch pairing, delete
account — has an app twin, so it goes, and a hand-off card points at
`${APP_URL}/profile/more`.

`aurora-credentials-section`'s two shared exports were lifted, not
deleted: `BoardCredentialCard` and `ImportProgressSteps` now live in
`components/board-entity/board-credential-card.tsx` next to their only
consumer, `board-import-prompt`, which the indexable `/aurora-migration`
and `/profile/{id}` still render.

No route is deleted, so no redirect, no locale twins and no `Disallow`
change: `/settings` survives and stays noindexed. Nothing here touches
`esp32_controllers` or its route handler.
