+++ packages/web/app/components/board-entity/board-credential-card.module.css
@@ -1,11 +1,10 @@
-.sectionDescription {
-  display: block;
-  margin-bottom: 16px;
-}
-
-.cardsContainer {
-  width: 100%;
-}
+/*
+ * Lifted from `components/settings/aurora-credentials-section.module.css` by
+ * W-21 (#4440), minus the three classes that only the deleted
+ * `AuroraCredentialsSection` used: `.sectionDescription`, `.cardsContainer` and
+ * `.loadingContainer`. Everything else is still bound by either
+ * `board-credential-card.tsx` or `board-import-prompt.tsx`.
+ */
 
 .credentialCard {
   width: 100%;
@@ -38,12 +37,6 @@
   padding: 8px;
   background-color: var(--color-error-bg);
   border-radius: 4px;
-}
-
-.loadingContainer {
-  display: flex;
-  justify-content: center;
-  padding: 24px;
 }
 
 .modalDescription {
