+++ packages/web/app/components/board-entity/board-import-prompt.tsx
@@ -33,14 +33,14 @@ import {
   ImportProgressSteps,
   type ImportPhase,
   type ImportProgress,
-} from '@/app/components/settings/aurora-credentials-section';
+} from './board-credential-card';
 import {
   parseAuroraExport,
   type AuroraExportPreview,
   type StrippedExportData,
 } from '@/app/lib/data-sync/aurora/parse-aurora-export';
 import type { AuroraBoardName } from '@boardsesh/shared-schema';
-import styles from '@/app/components/settings/aurora-credentials-section.module.css';
+import styles from './board-credential-card.module.css';
 
 type BoardImportPromptProps = {
   boardType: AuroraBoardName;
