--- packages/web/app/components/settings/aurora-credentials-section.tsx
@@ -1,1522 +1,0 @@
-'use client';
-
-import React, { useState, useEffect, useRef } from 'react';
-import Box from '@mui/material/Box';
-import Dialog from '@mui/material/Dialog';
-import DialogTitle from '@mui/material/DialogTitle';
-import DialogContent from '@mui/material/DialogContent';
-import DialogActions from '@mui/material/DialogActions';
-import MuiAlert from '@mui/material/Alert';
-import AlertTitle from '@mui/material/AlertTitle';
-import Chip from '@mui/material/Chip';
-import Stack from '@mui/material/Stack';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import ListItemText from '@mui/material/ListItemText';
-import LinearProgress from '@mui/material/LinearProgress';
-import { ConfirmPopover } from '@/app/components/ui/confirm-popover';
-import { useSnackbar } from '@/app/components/providers/snackbar-provider';
-import { LoadingSpinner } from '@/app/components/ui/loading-spinner';
-import { useWsAuthToken } from '@/app/hooks/use-ws-auth-token';
-import Typography from '@mui/material/Typography';
-import Button from '@mui/material/Button';
-import Card from '@mui/material/Card';
-import CardContent from '@mui/material/CardContent';
-import TextField from '@mui/material/TextField';
-import CircularProgress from '@mui/material/CircularProgress';
-import CheckCircleOutlined from '@mui/icons-material/CheckCircleOutlined';
-import WarningAmberOutlined from '@mui/icons-material/WarningAmberOutlined';
-import AccessTimeOutlined from '@mui/icons-material/AccessTimeOutlined';
-import DeleteOutlined from '@mui/icons-material/DeleteOutlined';
-import LinkOutlined from '@mui/icons-material/LinkOutlined';
-import SyncOutlined from '@mui/icons-material/SyncOutlined';
-import WarningOutlined from '@mui/icons-material/WarningOutlined';
-import EmailOutlined from '@mui/icons-material/EmailOutlined';
-import FileUploadOutlined from '@mui/icons-material/FileUploadOutlined';
-import RadioButtonUncheckedOutlined from '@mui/icons-material/RadioButtonUncheckedOutlined';
-import { useSession } from 'next-auth/react';
-import { usePathname, useRouter, useSearchParams } from 'next/navigation';
-import { Trans, useTranslation } from 'react-i18next';
-import type { TFunction } from 'i18next';
-import { DUPLICATE_BOARD_ACCOUNT_CIRCUITS_SYNC_ERROR } from '@boardsesh/shared-schema/sync-error-codes';
-import {
-  AuroraBackendError,
-  deleteAuroraCredential,
-  finalizeKilterCredential,
-  getAuroraCredentials,
-  getAuroraUnsyncedCounts,
-  resolveAuroraBackendTransport,
-  saveAuroraCredential,
-  saveKilterCredentialViaPassword,
-  type AuroraCredentialStatus,
-  type UnsyncedCounts,
-} from '@/app/lib/aurora-credentials/client';
-import { useFeatureFlag } from '@/app/components/providers/feature-flags-provider';
-import type { ImportResult } from '@/app/lib/data-sync/aurora/json-import';
-import { streamImport } from '@/app/lib/data-sync/aurora/json-import-stream';
-import {
-  parseAuroraExport,
-  type AuroraExportPreview,
-  type StrippedExportData,
-} from '@/app/lib/data-sync/aurora/parse-aurora-export';
-import {
-  parseMoonBoardExportCsvForImport,
-  streamMoonBoardImport,
-  type MoonBoardExportPreview,
-  type MoonBoardImportProgress as MoonBoardProgress,
-  type MoonBoardImportResult,
-  type StrippedMoonBoardExportData,
-} from '@/app/lib/data-sync/moonboard/csv-import-stream';
-import { AURORA_BOARDS, type AuroraBoardName } from '@boardsesh/shared-schema';
-import styles from './aurora-credentials-section.module.css';
-
-type BoardUnsyncedCounts = {
-  ascents: number;
-  climbs: number;
-};
-
-export type ImportPhase = 'preview' | 'importing' | 'complete' | 'error';
-
-export type ImportStep = 'climbs' | 'resolving' | 'dedup' | 'ascents' | 'attempts' | 'circuits';
-
-export type ImportProgress = {
-  step: ImportStep;
-  message?: string;
-  current?: number;
-  total?: number;
-};
-
-export const STEP_ORDER: ImportStep[] = ['climbs', 'resolving', 'dedup', 'ascents', 'attempts', 'circuits'];
-
-export const STEP_LABELS: Record<ImportStep, string> = {
-  climbs: 'Importing draft climbs',
-  resolving: 'Resolving climb names',
-  dedup: 'Checking for duplicates',
-  ascents: 'Importing ascents',
-  attempts: 'Importing attempts',
-  circuits: 'Importing circuits',
-};
-
-function getStepLabels(t: TFunction<'settings'>): Record<ImportStep, string> {
-  return {
-    climbs: t('aurora.import.steps.climbs'),
-    resolving: t('aurora.import.steps.resolving'),
-    dedup: t('aurora.import.steps.dedup'),
-    ascents: t('aurora.import.steps.ascents'),
-    attempts: t('aurora.import.steps.attempts'),
-    circuits: t('aurora.import.steps.circuits'),
-  };
-}
-
-function buildKilterDataRequestMailto(
-  t: TFunction<'settings'>,
-  userName?: string | null,
-  userEmail?: string | null,
-): string {
-  const name = userName || t('aurora.kilterEmail.namePlaceholder');
-  const email = userEmail || t('aurora.kilterEmail.emailPlaceholder');
-  const subject = t('aurora.kilterEmail.subject');
-  const body = t('aurora.kilterEmail.body', { name, email });
-
-  return `mailto:peter@auroraclimbing.com?subject=${encodeURIComponent(subject)}&body=${encodeURIComponent(body)}`;
-}
-
-const MOONBOARD_SUPPORT_EMAIL = 'moonboardsupport@moonclimbing.com';
-
-function buildMoonBoardDataRequestMailto(t: TFunction<'settings'>): string {
-  const subject = t('aurora.moonboard.email.subject');
-  return `mailto:${MOONBOARD_SUPPORT_EMAIL}?subject=${encodeURIComponent(subject)}`;
-}
-
-function totalMoonBoardImported(result: MoonBoardImportResult): number {
-  return result.ascents.imported + result.attempts.imported;
-}
-
-function moonBoardUnresolvedClimbs(result: MoonBoardImportResult): string[] {
-  return [
-    ...new Set([
-      ...result.unresolvedClimbs,
-      ...(result.unresolvedAscentClimbs ?? []),
-      ...(result.unresolvedAttemptClimbs ?? []),
-    ]),
-  ].sort();
-}
-
-function getMoonBoardProgressLabel(t: TFunction<'settings'>, progress: MoonBoardProgress | null): string {
-  if (!progress) return t('aurora.moonboard.csvImport.steps.resolving');
-  switch (progress.step) {
-    case 'ascents':
-      return t('aurora.moonboard.csvImport.steps.ascents');
-    case 'attempts':
-      return t('aurora.moonboard.csvImport.steps.attempts');
-    case 'dedup':
-      return t('aurora.moonboard.csvImport.steps.dedup');
-    case 'resolving':
-      return t('aurora.moonboard.csvImport.steps.resolving');
-    case 'importing':
-      return t('aurora.moonboard.csvImport.steps.importing');
-    case 'recomputing':
-      return t('aurora.moonboard.csvImport.steps.recomputing');
-    default:
-      return t('aurora.moonboard.csvImport.steps.importing');
-  }
-}
-
-function getMoonBoardImportErrorMessage(t: TFunction<'settings'>, error: unknown): string {
-  const errorCode = error instanceof Error ? error.message : typeof error === 'string' ? error : '';
-  if (errorCode === 'moonboard_import_interrupted') {
-    return t('aurora.moonboard.csvImport.interrupted');
-  }
-  return t('aurora.moonboard.csvImport.failed');
-}
-
-// Localise a link failure. The duplicate-account rejection carries a stable
-// `account_already_linked` code we map to a dedicated string; everything else
-// keeps the existing behaviour (backend message, or the generic fallback).
-// Both messages are passed pre-resolved so the i18n linter keeps its literal
-// `t('...')` calls at the call sites.
-function resolveLinkErrorMessage(error: unknown, alreadyLinkedMessage: string, fallbackMessage: string): string {
-  if (error instanceof AuroraBackendError && error.code === 'account_already_linked') {
-    return alreadyLinkedMessage;
-  }
-  return error instanceof Error ? error.message : fallbackMessage;
-}
-
-// Kilter renders two cards: `kilterAurora` for the legacy Aurora-built app (JSON
-// import + data request) and `kilterNew` for the new Kilter Grips account, which
-// links via username/password (ROPC). Every other board is a single `aurora` card.
-export type BoardCredentialCardVariant = 'aurora' | 'kilterAurora' | 'kilterNew';
-
-export type BoardCredentialCardProps = {
-  boardType: AuroraBoardName;
-  variant: BoardCredentialCardVariant;
-  credential: AuroraCredentialStatus | null;
-  unsyncedCounts: BoardUnsyncedCounts;
-  onAdd: () => void;
-  onRemove: () => void;
-  onImportJson: () => void;
-  isRemoving: boolean;
-  isImporting: boolean;
-  userName?: string | null;
-  userEmail?: string | null;
-};
-
-export function BoardCredentialCard({
-  boardType,
-  variant,
-  credential,
-  unsyncedCounts,
-  onAdd,
-  onRemove,
-  onImportJson,
-  isRemoving,
-  isImporting,
-  userName,
-  userEmail,
-}: BoardCredentialCardProps) {
-  const { t } = useTranslation('settings');
-  const boardName = boardType.charAt(0).toUpperCase() + boardType.slice(1);
-  const totalUnsynced = unsyncedCounts.ascents + unsyncedCounts.climbs;
-  const isExpired = credential?.syncStatus === 'expired';
-  const cardTitle =
-    variant === 'kilterAurora'
-      ? t('aurora.card.kilterAuroraTitle')
-      : variant === 'kilterNew'
-        ? t('aurora.card.kilterNewTitle')
-        : `${boardName} ${t('aurora.card.boardSuffix')}`;
-
-  const getSyncStatusTag = () => {
-    if (!credential) return null;
-
-    switch (credential.syncStatus) {
-      case 'active':
-        return (
-          <Chip icon={<CheckCircleOutlined />} label={t('aurora.status.connected')} size="small" color="success" />
-        );
-      case 'error':
-        return <Chip icon={<WarningAmberOutlined />} label={t('aurora.status.error')} size="small" color="error" />;
-      case 'expired':
-        return <Chip icon={<AccessTimeOutlined />} label={t('aurora.status.expired')} size="small" color="warning" />;
-      default:
-        return <Chip icon={<SyncOutlined />} label={t('aurora.status.syncing')} size="small" color="primary" />;
-    }
-  };
-
-  const formatLastSync = (dateString: string | null) => {
-    if (!dateString) return t('aurora.card.never');
-    const date = new Date(dateString);
-    return date.toLocaleDateString() + ' ' + date.toLocaleTimeString();
-  };
-
-  if (!credential) {
-    return (
-      <Card className={styles.credentialCard}>
-        <CardContent>
-          <div className={styles.cardHeader}>
-            <Typography variant="h5" sx={{ margin: 0 }}>
-              {cardTitle}
-            </Typography>
-          </div>
-          {variant === 'kilterNew' ? (
-            <>
-              <Typography variant="body2" component="span" color="text.secondary" className={styles.notConnectedText}>
-                {t('aurora.card.kilterNewCopy')}
-              </Typography>
-              <div className={styles.buttonRowStacked}>
-                <Button variant="contained" startIcon={<LinkOutlined />} onClick={onAdd}>
-                  {t('aurora.card.kilterSignIn')}
-                </Button>
-                <Button
-                  variant="outlined"
-                  startIcon={isImporting ? <CircularProgress size={16} /> : <FileUploadOutlined />}
-                  onClick={onImportJson}
-                  disabled={isImporting}
-                >
-                  {t('aurora.card.import')}
-                </Button>
-              </div>
-            </>
-          ) : variant === 'kilterAurora' ? (
-            <>
-              <Typography variant="body2" component="span" color="text.secondary" className={styles.notConnectedText}>
-                {t('aurora.card.kilterAuroraCopy')}
-              </Typography>
-              <div className={styles.buttonRowStacked}>
-                <Button
-                  variant="contained"
-                  startIcon={isImporting ? <CircularProgress size={16} /> : <FileUploadOutlined />}
-                  onClick={onImportJson}
-                  disabled={isImporting}
-                >
-                  {t('aurora.card.import')}
-                </Button>
-                <Button
-                  variant="outlined"
-                  startIcon={<EmailOutlined />}
-                  href={buildKilterDataRequestMailto(t, userName, userEmail)}
-                >
-                  {t('aurora.card.requestData')}
-                </Button>
-              </div>
-            </>
-          ) : (
-            <>
-              <Typography variant="body2" component="span" color="text.secondary" className={styles.notConnectedText}>
-                {t('aurora.card.notConnected', { boardName })}
-              </Typography>
-              <div className={styles.buttonRow}>
-                <Button variant="contained" startIcon={<LinkOutlined />} onClick={onAdd}>
-                  {t('aurora.card.link')}
-                </Button>
-                <Button
-                  variant="outlined"
-                  startIcon={isImporting ? <CircularProgress size={16} /> : <FileUploadOutlined />}
-                  onClick={onImportJson}
-                  disabled={isImporting}
-                >
-                  {t('aurora.card.import')}
-                </Button>
-              </div>
-            </>
-          )}
-        </CardContent>
-      </Card>
-    );
-  }
-
-  return (
-    <Card className={styles.credentialCard}>
-      <CardContent>
-        <div className={styles.cardHeader}>
-          <Typography variant="h5" sx={{ margin: 0 }}>
-            {cardTitle}
-          </Typography>
-          {getSyncStatusTag()}
-        </div>
-        <div className={styles.credentialInfo}>
-          <div className={styles.infoRow}>
-            <Typography variant="body2" component="span" color="text.secondary">
-              {t('aurora.card.username')}
-            </Typography>
-            <Typography variant="body2" component="span" fontWeight={600}>
-              {credential.auroraUsername}
-            </Typography>
-          </div>
-          <div className={styles.infoRow}>
-            <Typography variant="body2" component="span" color="text.secondary">
-              {t('aurora.card.lastSynced')}
-            </Typography>
-            <Typography variant="body2" component="span">
-              {formatLastSync(credential.lastSyncAt)}
-            </Typography>
-          </div>
-          {credential.syncError && (
-            <div className={styles.errorRow}>
-              {/* Known codes get localised copy and a warning tone — the account is
-                  syncing, only its playlist mirror is paused (#3526). Anything else
-                  is legacy free text from an older path: render it verbatim rather
-                  than swallow it. */}
-              <Typography
-                variant="body2"
-                component="span"
-                color={credential.syncError === DUPLICATE_BOARD_ACCOUNT_CIRCUITS_SYNC_ERROR ? 'warning.main' : 'error'}
-              >
-                {credential.syncError === DUPLICATE_BOARD_ACCOUNT_CIRCUITS_SYNC_ERROR
-                  ? t('aurora.status.duplicateAccountCircuits')
-                  : credential.syncError}
-              </Typography>
-            </div>
-          )}
-          {totalUnsynced > 0 && (
-            <MuiAlert severity="warning" icon={<WarningOutlined />} className={styles.unsyncedAlert}>
-              <AlertTitle>{t('aurora.unsynced.title', { count: totalUnsynced })}</AlertTitle>
-              <Typography variant="body2" component="span" color="text.secondary">
-                {unsyncedCounts.ascents > 0 && t('aurora.unsynced.ascents', { count: unsyncedCounts.ascents })}
-                {unsyncedCounts.ascents > 0 && unsyncedCounts.climbs > 0 && ', '}
-                {unsyncedCounts.climbs > 0 && t('aurora.unsynced.climbs', { count: unsyncedCounts.climbs })}
-              </Typography>
-            </MuiAlert>
-          )}
-        </div>
-        <div className={styles.buttonRow}>
-          {isExpired && (
-            <Button variant="contained" startIcon={<LinkOutlined />} onClick={onAdd}>
-              {t('aurora.card.reconnect')}
-            </Button>
-          )}
-          <ConfirmPopover
-            title={t('aurora.card.unlinkConfirm.title')}
-            description={t('aurora.card.unlinkConfirm.description', { boardName })}
-            onConfirm={onRemove}
-            okText={t('aurora.card.unlinkConfirm.ok')}
-            okButtonProps={{ color: 'error' }}
-          >
-            <Button
-              color="error"
-              variant="outlined"
-              startIcon={isRemoving ? <CircularProgress size={16} /> : <DeleteOutlined />}
-              disabled={isRemoving}
-            >
-              {t('aurora.card.unlink')}
-            </Button>
-          </ConfirmPopover>
-          <Button
-            variant="outlined"
-            startIcon={isImporting ? <CircularProgress size={16} /> : <FileUploadOutlined />}
-            onClick={onImportJson}
-            disabled={isImporting}
-          >
-            {t('aurora.card.import')}
-          </Button>
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
-
-type MoonBoardCredentialCardProps = {
-  onImportCsv: () => void;
-  onRequestData: () => void;
-  requestDataHref: string;
-  isImporting: boolean;
-};
-
-function MoonBoardCredentialCard({
-  onImportCsv,
-  onRequestData,
-  requestDataHref,
-  isImporting,
-}: MoonBoardCredentialCardProps) {
-  const { t } = useTranslation('settings');
-
-  return (
-    <Card className={styles.credentialCard}>
-      <CardContent>
-        <div className={styles.cardHeader}>
-          <Typography variant="h5" sx={{ margin: 0 }}>
-            {t('aurora.moonboard.title')}
-          </Typography>
-        </div>
-        <Typography variant="body2" component="span" color="text.secondary" className={styles.notConnectedText}>
-          {t('aurora.moonboard.copy')}
-        </Typography>
-        <div className={styles.buttonRowStacked}>
-          <Button
-            variant="contained"
-            startIcon={isImporting ? <CircularProgress size={16} /> : <FileUploadOutlined />}
-            onClick={onImportCsv}
-            disabled={isImporting}
-          >
-            {t('aurora.moonboard.import')}
-          </Button>
-          <Button variant="outlined" startIcon={<EmailOutlined />} href={requestDataHref} onClick={onRequestData}>
-            {t('aurora.moonboard.requestData')}
-          </Button>
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
-
-export function ImportProgressSteps({ progress }: { progress: ImportProgress | null }) {
-  const { t } = useTranslation('settings');
-  const stepLabels = getStepLabels(t);
-  const currentStepIndex = progress ? STEP_ORDER.indexOf(progress.step) : -1;
-
-  return (
-    <div className={styles.progressStepList}>
-      {STEP_ORDER.map((step, index) => {
-        const isComplete = index < currentStepIndex;
-        const isActive = index === currentStepIndex;
-        const isPending = index > currentStepIndex;
-        const hasCounts = isActive && progress?.current != null && progress?.total != null;
-        const progressPercent = hasCounts ? (progress.current! / progress.total!) * 100 : 0;
-
-        return (
-          <div key={step} className={styles.progressStep}>
-            <div className={styles.progressStepHeader}>
-              {isComplete && <CheckCircleOutlined color="success" fontSize="small" />}
-              {isActive && <CircularProgress size={20} />}
-              {isPending && <RadioButtonUncheckedOutlined color="disabled" fontSize="small" />}
-              <Typography
-                variant="body2"
-                color={isPending ? 'text.disabled' : 'text.primary'}
-                fontWeight={isActive ? 600 : 400}
-              >
-                {stepLabels[step]}
-                {hasCounts && ` (${progress.current} / ${progress.total})`}
-              </Typography>
-            </div>
-            {hasCounts && (
-              <LinearProgress variant="determinate" value={progressPercent} sx={{ ml: '32px', mr: 1, mt: 0.5 }} />
-            )}
-            {isActive && !hasCounts && <LinearProgress variant="indeterminate" sx={{ ml: '32px', mr: 1, mt: 0.5 }} />}
-          </div>
-        );
-      })}
-    </div>
-  );
-}
-
-export default function AuroraCredentialsSection() {
-  const { t } = useTranslation('settings');
-  const { data: session } = useSession();
-  const router = useRouter();
-  const pathname = usePathname();
-  const searchParams = useSearchParams();
-  const { token: authToken, isLoading: authTokenLoading } = useWsAuthToken();
-  const { showMessage } = useSnackbar();
-  const [credentials, setCredentials] = useState<AuroraCredentialStatus[]>([]);
-  const [unsyncedCounts, setUnsyncedCounts] = useState<UnsyncedCounts | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [credentialsLoadError, setCredentialsLoadError] = useState(false);
-  const kilterOauthLinkingEnabled = useFeatureFlag('kilter-oauth-linking') === true;
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [selectedBoard, setSelectedBoard] = useState<AuroraBoardName>('kilter');
-  const [isSaving, setIsSaving] = useState(false);
-  const [removingBoard, setRemovingBoard] = useState<string | null>(null);
-  const [formValues, setFormValues] = useState({ username: '', password: '' });
-
-  // Import state
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const [importingBoard, setImportingBoard] = useState<AuroraBoardName | null>(null);
-  const [importPreview, setAuroraExportPreview] = useState<AuroraExportPreview | null>(null);
-  const [importRawData, setImportRawData] = useState<StrippedExportData | null>(null);
-  const [importPhase, setImportPhase] = useState<ImportPhase | null>(null);
-  const [importProgress, setImportProgress] = useState<ImportProgress | null>(null);
-  const [importResult, setImportResult] = useState<ImportResult | null>(null);
-  const [importError, setImportError] = useState<string | null>(null);
-  const receivedCompleteRef = useRef(false);
-  const handledKilterCompletionRef = useRef<string | null>(null);
-  const moonBoardFileInputRef = useRef<HTMLInputElement>(null);
-  const [moonBoardImportPreview, setMoonBoardImportPreview] = useState<MoonBoardExportPreview | null>(null);
-  const [moonBoardImportData, setMoonBoardImportData] = useState<StrippedMoonBoardExportData | null>(null);
-  const [moonBoardImportPhase, setMoonBoardImportPhase] = useState<ImportPhase | null>(null);
-  const [moonBoardImportProgress, setMoonBoardImportProgress] = useState<MoonBoardProgress | null>(null);
-  const [moonBoardImportResult, setMoonBoardImportResult] = useState<MoonBoardImportResult | null>(null);
-  const [moonBoardImportError, setMoonBoardImportError] = useState<string | null>(null);
-  const receivedMoonBoardCompleteRef = useRef(false);
-
-  const fetchCredentials = async () => {
-    setCredentialsLoadError(false);
-    const transport = resolveAuroraBackendTransport(authToken);
-    if (!transport) {
-      setCredentialsLoadError(true);
-      setLoading(false);
-      return;
-    }
-
-    try {
-      const data = await getAuroraCredentials(transport);
-      setCredentials(data.credentials);
-    } catch (error) {
-      console.error('Failed to fetch credentials:', error);
-      setCredentialsLoadError(true);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const fetchUnsyncedCounts = async () => {
-    const transport = resolveAuroraBackendTransport(authToken);
-    if (!transport) return;
-
-    try {
-      setUnsyncedCounts(await getAuroraUnsyncedCounts(transport));
-    } catch (error) {
-      console.error('Failed to fetch unsynced counts:', error);
-    }
-  };
-
-  useEffect(() => {
-    if (authTokenLoading) return;
-    setLoading(true);
-    void fetchCredentials();
-    void fetchUnsyncedCounts();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [authToken, authTokenLoading]);
-
-  const clearKilterSearchParams = () => {
-    const nextParams = new URLSearchParams(searchParams.toString());
-    nextParams.delete('kilter');
-    nextParams.delete('reason');
-    nextParams.delete('completion');
-    const query = nextParams.toString();
-    router.replace(query ? `${pathname}?${query}` : pathname, { scroll: false });
-  };
-
-  useEffect(() => {
-    if (authTokenLoading) return;
-
-    const kilterStatus = searchParams.get('kilter');
-    if (!kilterStatus) return;
-
-    if (kilterStatus === 'error') {
-      // The OAuth callback can carry `reason=account_already_linked` when the
-      // browser flow (not the /finalize POST) hits a duplicate-account link.
-      // Route it through the same helper so it gets the dedicated copy instead
-      // of the generic failure string.
-      const reason = searchParams.get('reason');
-      const linkError =
-        reason === 'account_already_linked'
-          ? new AuroraBackendError(409, 'account_already_linked', 'account_already_linked')
-          : undefined;
-      showMessage(
-        resolveLinkErrorMessage(
-          linkError,
-          t('aurora.linkDialog.accountAlreadyLinked'),
-          t('aurora.mobile.kilterConnectFailed'),
-        ),
-        'error',
-      );
-      clearKilterSearchParams();
-      return;
-    }
-
-    if (kilterStatus !== 'connected') return;
-
-    const completion = searchParams.get('completion');
-    if (!completion) {
-      showMessage(t('aurora.mobile.kilterConnectFailed'), 'error');
-      clearKilterSearchParams();
-      return;
-    }
-    if (handledKilterCompletionRef.current === completion) return;
-    handledKilterCompletionRef.current = completion;
-
-    void (async () => {
-      try {
-        const transport = resolveAuroraBackendTransport(authToken);
-        if (!transport) throw new Error(t('aurora.mobile.kilterConnectFailed'));
-
-        await finalizeKilterCredential(transport, completion);
-        showMessage(t('aurora.mobile.kilterConnected'), 'success');
-        await Promise.all([fetchCredentials(), fetchUnsyncedCounts()]);
-      } catch (error) {
-        handledKilterCompletionRef.current = null;
-        showMessage(
-          resolveLinkErrorMessage(
-            error,
-            t('aurora.linkDialog.accountAlreadyLinked'),
-            t('aurora.mobile.kilterConnectFailed'),
-          ),
-          'error',
-        );
-      } finally {
-        clearKilterSearchParams();
-      }
-    })();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [authToken, authTokenLoading, searchParams]);
-
-  const handleAddClick = (boardType: AuroraBoardName) => {
-    setSelectedBoard(boardType);
-    setFormValues({ username: '', password: '' });
-    setIsModalOpen(true);
-  };
-
-  const handleModalCancel = () => {
-    setIsModalOpen(false);
-    setFormValues({ username: '', password: '' });
-  };
-
-  const selectedBoardName = selectedBoard.charAt(0).toUpperCase() + selectedBoard.slice(1);
-
-  const handleSaveCredentials = async (values: { username: string; password: string }) => {
-    setIsSaving(true);
-    try {
-      const transport = resolveAuroraBackendTransport(authToken);
-      if (!transport) throw new Error(t('aurora.linkDialog.linkError'));
-
-      if (selectedBoard === 'kilter') {
-        await saveKilterCredentialViaPassword(transport, {
-          username: values.username,
-          password: values.password,
-        });
-      } else {
-        await saveAuroraCredential(transport, {
-          boardType: selectedBoard,
-          username: values.username,
-          password: values.password,
-        });
-      }
-
-      showMessage(t('aurora.linkDialog.linkSuccess', { boardName: selectedBoardName }), 'success');
-      setIsModalOpen(false);
-      setFormValues({ username: '', password: '' });
-      await fetchCredentials();
-    } catch (error) {
-      showMessage(
-        resolveLinkErrorMessage(error, t('aurora.linkDialog.accountAlreadyLinked'), t('aurora.linkDialog.linkError')),
-        'error',
-      );
-    } finally {
-      setIsSaving(false);
-    }
-  };
-
-  const handleRemove = async (boardType: AuroraBoardName) => {
-    setRemovingBoard(boardType);
-    try {
-      const transport = resolveAuroraBackendTransport(authToken);
-      if (!transport) throw new Error(t('aurora.unlinkError'));
-
-      const result = await deleteAuroraCredential(transport, boardType);
-
-      if (!result.success && result.reason === 'revocation_failed') {
-        showMessage(t('aurora.unlinkPartial'), 'warning');
-      } else {
-        showMessage(t('aurora.unlinkSuccess'), 'success');
-      }
-      await fetchCredentials();
-    } catch (error) {
-      showMessage(error instanceof Error ? error.message : t('aurora.unlinkError'), 'error');
-    } finally {
-      setRemovingBoard(null);
-    }
-  };
-
-  // --- JSON Import handlers ---
-
-  const resetImportState = () => {
-    setImportPhase(null);
-    setImportProgress(null);
-    setAuroraExportPreview(null);
-    setImportRawData(null);
-    setImportingBoard(null);
-    setImportResult(null);
-    setImportError(null);
-  };
-
-  const handleImportClick = (boardType: AuroraBoardName) => {
-    setImportingBoard(boardType);
-    fileInputRef.current?.click();
-  };
-
-  const handleFileSelected = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0];
-    if (!file || !importingBoard) return;
-
-    // Guard against very large files (200MB limit - exports can be large due to climb data)
-    const maxSizeBytes = 200 * 1024 * 1024;
-    if (file.size > maxSizeBytes) {
-      showMessage(t('aurora.import.tooLarge'), 'error');
-      setImportingBoard(null);
-      event.target.value = '';
-      return;
-    }
-
-    const reader = new FileReader();
-    reader.onload = (e) => {
-      try {
-        const json = JSON.parse(e.target?.result as string);
-        const parsed = parseAuroraExport(json, importingBoard);
-
-        if (parsed.boardWarning) {
-          showMessage(parsed.boardWarning, 'warning');
-        }
-
-        setImportRawData(parsed.data);
-        setAuroraExportPreview(parsed.preview);
-        setImportPhase('preview');
-      } catch (err) {
-        showMessage(err instanceof Error ? err.message : t('aurora.import.parseError'), 'error');
-        setImportingBoard(null);
-      }
-    };
-    reader.onerror = () => {
-      showMessage(t('aurora.import.readError'), 'error');
-      setImportingBoard(null);
-    };
-    reader.readAsText(file);
-
-    // Reset input so the same file can be re-selected
-    event.target.value = '';
-  };
-
-  const handleImportConfirm = async () => {
-    if (!importingBoard || !importRawData) return;
-
-    setImportPhase('importing');
-    setImportProgress(null);
-    setAuroraExportPreview(null);
-    receivedCompleteRef.current = false;
-
-    try {
-      const transport = resolveAuroraBackendTransport(authToken);
-      if (!transport) throw new Error(t('aurora.import.failed'));
-
-      await streamImport(
-        importingBoard,
-        importRawData,
-        (event) => {
-          switch (event.type) {
-            case 'progress':
-              setImportProgress({
-                step: event.step,
-                message: 'message' in event ? event.message : undefined,
-                current: 'current' in event ? event.current : undefined,
-                total: 'total' in event ? event.total : undefined,
-              });
-              break;
-            case 'complete':
-              receivedCompleteRef.current = true;
-              setImportResult(event.results);
-              setImportPhase('complete');
-              {
-                const totalImported =
-                  event.results.climbs.imported +
-                  event.results.ascents.imported +
-                  event.results.attempts.imported +
-                  event.results.circuits.imported;
-                if (event.results.partialError) {
-                  showMessage(t('aurora.import.partialWarning'), 'warning');
-                } else {
-                  showMessage(t('aurora.import.successCount', { count: totalImported }), 'success');
-                }
-              }
-              break;
-            case 'error':
-              receivedCompleteRef.current = true;
-              setImportError(event.error);
-              setImportPhase('error');
-              showMessage(event.error, 'error');
-              break;
-          }
-        },
-        {
-          backendUrl: transport.backendUrl,
-          authToken: transport.authToken,
-        },
-      );
-
-      // If stream ended without a complete/error event (e.g. server timeout)
-      if (!receivedCompleteRef.current) {
-        setImportError(t('aurora.import.interrupted'));
-        setImportPhase('error');
-        showMessage(t('aurora.import.interruptedShort'), 'error');
-      }
-    } catch (error) {
-      const message = error instanceof Error ? error.message : t('aurora.import.failed');
-      setImportError(message);
-      setImportPhase('error');
-      showMessage(message, 'error');
-    } finally {
-      setImportRawData(null);
-    }
-  };
-
-  const handleImportDialogClose = () => {
-    if (importPhase === 'importing') return;
-    resetImportState();
-  };
-
-  const resetMoonBoardImportState = () => {
-    setMoonBoardImportPhase(null);
-    setMoonBoardImportProgress(null);
-    setMoonBoardImportPreview(null);
-    setMoonBoardImportData(null);
-    setMoonBoardImportResult(null);
-    setMoonBoardImportError(null);
-  };
-
-  const handleMoonBoardImportClick = () => {
-    moonBoardFileInputRef.current?.click();
-  };
-
-  const handleMoonBoardRequestDataClick = () => {
-    if (!navigator.clipboard) return;
-
-    void navigator.clipboard.writeText(t('aurora.moonboard.email.body')).then(
-      () => showMessage(t('aurora.moonboard.requestDataDialog.title'), 'success'),
-      () => showMessage(t('aurora.mobile.requestDataCopyFailed'), 'error'),
-    );
-  };
-
-  const handleMoonBoardFileSelected = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0];
-    if (!file) return;
-
-    const maxSizeBytes = 200 * 1024 * 1024;
-    if (file.size > maxSizeBytes) {
-      showMessage(t('aurora.import.tooLarge'), 'error');
-      event.target.value = '';
-      return;
-    }
-
-    const reader = new FileReader();
-    reader.onload = () => {
-      void (async () => {
-        try {
-          const csv = typeof reader.result === 'string' ? reader.result : '';
-          const parsed = await parseMoonBoardExportCsvForImport(csv);
-          setMoonBoardImportData(parsed.data);
-          setMoonBoardImportPreview(parsed.preview);
-          setMoonBoardImportPhase('preview');
-        } catch (error) {
-          const message =
-            error instanceof Error && error.message === 'moonboard_parser_unavailable'
-              ? t('aurora.moonboard.csvImport.parserUnavailable')
-              : t('aurora.moonboard.csvImport.parseError');
-          showMessage(message, 'error');
-          resetMoonBoardImportState();
-        }
-      })();
-    };
-    reader.onerror = () => {
-      showMessage(t('aurora.import.readError'), 'error');
-      resetMoonBoardImportState();
-    };
-    reader.readAsText(file);
-
-    event.target.value = '';
-  };
-
-  const handleMoonBoardImportConfirm = async () => {
-    if (!moonBoardImportData) return;
-
-    setMoonBoardImportPhase('importing');
-    setMoonBoardImportProgress(null);
-    setMoonBoardImportPreview(null);
-    setMoonBoardImportResult(null);
-    setMoonBoardImportError(null);
-    receivedMoonBoardCompleteRef.current = false;
-
-    try {
-      const transport = resolveAuroraBackendTransport(authToken);
-      if (!transport) throw new Error('moonboard_import_endpoint_unavailable');
-
-      await streamMoonBoardImport(
-        moonBoardImportData,
-        (event) => {
-          switch (event.type) {
-            case 'progress':
-              setMoonBoardImportProgress({
-                step: event.step,
-                message: 'message' in event ? event.message : undefined,
-                current: 'current' in event ? event.current : undefined,
-                total: 'total' in event ? event.total : undefined,
-              });
-              break;
-            case 'complete':
-              receivedMoonBoardCompleteRef.current = true;
-              setMoonBoardImportResult(event.results);
-              setMoonBoardImportPhase('complete');
-              showMessage(
-                event.results.partialError
-                  ? t('aurora.moonboard.csvImport.partialWarning')
-                  : t('aurora.moonboard.csvImport.successCount', { count: totalMoonBoardImported(event.results) }),
-                event.results.partialError ? 'warning' : 'success',
-              );
-              break;
-            case 'error': {
-              receivedMoonBoardCompleteRef.current = true;
-              const importErrorMessage = getMoonBoardImportErrorMessage(t, event.error);
-              setMoonBoardImportError(importErrorMessage);
-              setMoonBoardImportPhase('error');
-              showMessage(importErrorMessage, 'error');
-              break;
-            }
-          }
-        },
-        {
-          backendUrl: transport.backendUrl,
-          authToken: transport.authToken,
-        },
-      );
-
-      if (!receivedMoonBoardCompleteRef.current) {
-        setMoonBoardImportError(t('aurora.moonboard.csvImport.interrupted'));
-        setMoonBoardImportPhase('error');
-        showMessage(t('aurora.moonboard.csvImport.interruptedShort'), 'error');
-      }
-    } catch (error) {
-      const message = getMoonBoardImportErrorMessage(t, error);
-      setMoonBoardImportError(message);
-      setMoonBoardImportPhase('error');
-      showMessage(message, 'error');
-    } finally {
-      setMoonBoardImportData(null);
-    }
-  };
-
-  const handleMoonBoardImportDialogClose = () => {
-    if (moonBoardImportPhase === 'importing') return;
-    resetMoonBoardImportState();
-  };
-
-  const getCredentialForBoard = (boardType: AuroraBoardName) => {
-    return credentials.find((credential) => credential.boardType === boardType) || null;
-  };
-
-  // Show the new Kilter card when the `kilter-oauth-linking` flag is on, or whenever a
-  // Kilter account is already linked (so it stays manageable if the flag flips off).
-  const showKilterNew = kilterOauthLinkingEnabled || getCredentialForBoard('kilter') !== null;
-  const cardConfigs = AURORA_BOARDS.flatMap<{
-    key: string;
-    boardType: AuroraBoardName;
-    variant: BoardCredentialCardVariant;
-  }>((boardType) => {
-    if (boardType !== 'kilter') return [{ key: boardType, boardType, variant: 'aurora' }];
-    const kilterCards: { key: string; boardType: AuroraBoardName; variant: BoardCredentialCardVariant }[] = [
-      { key: 'kilter-aurora', boardType: 'kilter', variant: 'kilterAurora' },
-    ];
-    if (showKilterNew) kilterCards.push({ key: 'kilter-new', boardType: 'kilter', variant: 'kilterNew' });
-    return kilterCards;
-  });
-
-  const isImporting = importPhase === 'importing';
-  const isImportDialogOpen =
-    importPhase === 'preview' || importPhase === 'importing' || importPhase === 'complete' || importPhase === 'error';
-  const isMoonBoardImporting = moonBoardImportPhase === 'importing';
-  const isMoonBoardImportDialogOpen =
-    moonBoardImportPhase === 'preview' ||
-    moonBoardImportPhase === 'importing' ||
-    moonBoardImportPhase === 'complete' ||
-    moonBoardImportPhase === 'error';
-
-  const getImportDialogTitle = () => {
-    switch (importPhase) {
-      case 'preview':
-        return t('aurora.import.dialog.previewTitle');
-      case 'importing':
-        return t('aurora.import.dialog.importingTitle');
-      case 'complete':
-        return t('aurora.import.dialog.completeTitle');
-      case 'error':
-        return t('aurora.import.dialog.errorTitle');
-      default:
-        return '';
-    }
-  };
-
-  const getMoonBoardImportDialogTitle = () => {
-    switch (moonBoardImportPhase) {
-      case 'preview':
-        return t('aurora.moonboard.importDialog.previewTitle');
-      case 'importing':
-        return t('aurora.moonboard.importDialog.importingTitle');
-      case 'complete':
-        return t('aurora.moonboard.importDialog.completeTitle');
-      case 'error':
-        return t('aurora.moonboard.importDialog.errorTitle');
-      default:
-        return '';
-    }
-  };
-
-  const importingBoardName = importingBoard ? importingBoard.charAt(0).toUpperCase() + importingBoard.slice(1) : '';
-  const moonBoardProgressLabel = getMoonBoardProgressLabel(t, moonBoardImportProgress);
-  const moonBoardUnresolvedNames = moonBoardImportResult ? moonBoardUnresolvedClimbs(moonBoardImportResult) : [];
-
-  if (loading) {
-    return (
-      <Card>
-        <CardContent>
-          <div className={styles.loadingContainer}>
-            <LoadingSpinner />
-          </div>
-        </CardContent>
-      </Card>
-    );
-  }
-
-  if (credentialsLoadError) {
-    return (
-      <Card>
-        <CardContent>
-          <MuiAlert
-            severity="error"
-            action={
-              <Button
-                color="inherit"
-                size="small"
-                onClick={() => {
-                  setLoading(true);
-                  void fetchCredentials();
-                  void fetchUnsyncedCounts();
-                }}
-              >
-                {t('aurora.loadRetry')}
-              </Button>
-            }
-          >
-            {t('aurora.mobile.loadFailed')}
-          </MuiAlert>
-        </CardContent>
-      </Card>
-    );
-  }
-
-  return (
-    <>
-      <Card>
-        <CardContent>
-          <Typography variant="h5">{t('aurora.title')}</Typography>
-          <Typography variant="body2" component="span" color="text.secondary" className={styles.sectionDescription}>
-            {t('aurora.subtitle')}
-          </Typography>
-
-          <Stack spacing={2} className={styles.cardsContainer}>
-            <MoonBoardCredentialCard
-              onImportCsv={handleMoonBoardImportClick}
-              onRequestData={handleMoonBoardRequestDataClick}
-              requestDataHref={buildMoonBoardDataRequestMailto(t)}
-              isImporting={isMoonBoardImporting}
-            />
-            {cardConfigs.map(({ key, boardType, variant }) => {
-              // The legacy "Kilter (Aurora)" card never owns the linked-account state —
-              // that belongs to the new card.
-              const credential = variant === 'kilterAurora' ? null : getCredentialForBoard(boardType);
-              return (
-                <BoardCredentialCard
-                  key={key}
-                  boardType={boardType}
-                  variant={variant}
-                  credential={credential}
-                  unsyncedCounts={unsyncedCounts?.[boardType] ?? { ascents: 0, climbs: 0 }}
-                  onAdd={() => handleAddClick(boardType)}
-                  onRemove={() => handleRemove(boardType)}
-                  onImportJson={() => handleImportClick(boardType)}
-                  isRemoving={removingBoard === boardType}
-                  isImporting={isImporting && importingBoard === boardType}
-                  userName={boardType === 'kilter' ? (session?.user?.name ?? null) : null}
-                  userEmail={boardType === 'kilter' ? (session?.user?.email ?? null) : null}
-                />
-              );
-            })}
-          </Stack>
-        </CardContent>
-      </Card>
-
-      {/* Hidden file input for JSON import */}
-      <input
-        ref={fileInputRef}
-        type="file"
-        accept=".json"
-        onChange={handleFileSelected}
-        aria-label={t('aurora.import.fileLabel')}
-        hidden
-      />
-
-      {/* Hidden file input for MoonBoard CSV import */}
-      <input
-        ref={moonBoardFileInputRef}
-        type="file"
-        accept=".csv,text/csv"
-        onChange={handleMoonBoardFileSelected}
-        aria-label={t('aurora.moonboard.csvImport.fileLabel')}
-        hidden
-      />
-
-      {/* Link Account Dialog */}
-      <Dialog open={isModalOpen} onClose={handleModalCancel} maxWidth="sm" fullWidth>
-        <DialogTitle>
-          {selectedBoard === 'kilter'
-            ? t('aurora.kilterLinkDialog.title')
-            : t('aurora.linkDialog.title', { boardName: selectedBoardName })}
-        </DialogTitle>
-        <DialogContent>
-          <Typography variant="body2" component="span" color="text.secondary" className={styles.modalDescription}>
-            {selectedBoard === 'kilter'
-              ? t('aurora.kilterLinkDialog.description')
-              : t('aurora.linkDialog.description', { boardName: selectedBoardName })}
-          </Typography>
-          <Box
-            component="form"
-            onSubmit={(e: React.FormEvent) => {
-              e.preventDefault();
-              const vals = formValues;
-              if (!vals.username || !vals.password) return;
-              void handleSaveCredentials(vals);
-            }}
-            sx={{ display: 'flex', flexDirection: 'column', gap: 2, mt: 2 }}
-          >
-            <TextField
-              label={t('aurora.linkDialog.usernameLabel')}
-              placeholder={t('aurora.linkDialog.usernamePlaceholder')}
-              variant="outlined"
-              size="small"
-              fullWidth
-              required
-              value={formValues.username}
-              onChange={(e) => setFormValues((prev) => ({ ...prev, username: e.target.value }))}
-            />
-
-            <TextField
-              label={t('aurora.linkDialog.passwordLabel')}
-              type="password"
-              placeholder={t('aurora.linkDialog.passwordPlaceholder')}
-              variant="outlined"
-              size="small"
-              fullWidth
-              required
-              value={formValues.password}
-              onChange={(e) => setFormValues((prev) => ({ ...prev, password: e.target.value }))}
-            />
-
-            <Button
-              variant="contained"
-              type="submit"
-              disabled={isSaving}
-              startIcon={isSaving ? <CircularProgress size={16} /> : undefined}
-              fullWidth
-            >
-              {isSaving ? t('aurora.linkDialog.submitting') : t('aurora.linkDialog.submit')}
-            </Button>
-          </Box>
-        </DialogContent>
-      </Dialog>
-
-      {/* Unified Import Dialog (preview → progress → complete/error) */}
-      <Dialog
-        open={isImportDialogOpen}
-        onClose={handleImportDialogClose}
-        maxWidth="sm"
-        fullWidth
-        disableEscapeKeyDown={isImporting}
-      >
-        <DialogTitle>{getImportDialogTitle()}</DialogTitle>
-        <DialogContent>
-          {/* Preview phase */}
-          {importPhase === 'preview' && importPreview && (
-            <>
-              <Typography variant="body2" color="text.secondary" className={styles.modalDescription}>
-                <Trans
-                  i18nKey="aurora.import.dialog.previewIntro"
-                  t={t}
-                  values={{ username: importPreview.username, boardName: importingBoardName }}
-                  components={{ strong: <strong /> }}
-                />
-              </Typography>
-              <List dense>
-                {importPreview.climbs > 0 && (
-                  <ListItem>
-                    <ListItemText primary={t('aurora.import.dialog.draftClimbs', { count: importPreview.climbs })} />
-                  </ListItem>
-                )}
-                <ListItem>
-                  <ListItemText primary={t('aurora.import.dialog.ascents', { count: importPreview.ascents })} />
-                </ListItem>
-                <ListItem>
-                  <ListItemText primary={t('aurora.import.dialog.attempts', { count: importPreview.attempts })} />
-                </ListItem>
-                <ListItem>
-                  <ListItemText primary={t('aurora.import.dialog.circuits', { count: importPreview.circuits })} />
-                </ListItem>
-              </List>
-              <Typography variant="body2" color="text.secondary">
-                {t('aurora.import.dialog.previewNote')}
-              </Typography>
-            </>
-          )}
-
-          {/* Importing phase */}
-          {importPhase === 'importing' && <ImportProgressSteps progress={importProgress} />}
-
-          {/* Complete phase */}
-          {importPhase === 'complete' && importResult && (
-            <>
-              {importResult.partialError && (
-                <MuiAlert severity="warning" className={styles.unsyncedAlert}>
-                  <AlertTitle>{t('aurora.import.results.partialTitle')}</AlertTitle>
-                  {t('aurora.import.results.partialBody')}
-                </MuiAlert>
-              )}
-              <List dense>
-                {(importResult.climbs.imported > 0 || importResult.climbs.failed > 0) && (
-                  <ListItem>
-                    <ListItemText
-                      primary={t('aurora.import.results.draftClimbs')}
-                      secondary={t('aurora.import.results.draftClimbsSummary', {
-                        imported: importResult.climbs.imported,
-                        skipped: importResult.climbs.skipped,
-                        failed: importResult.climbs.failed,
-                      })}
-                    />
-                  </ListItem>
-                )}
-                <ListItem>
-                  <ListItemText
-                    primary={t('aurora.import.results.ascents')}
-                    secondary={t('aurora.import.results.ascentsSummary', {
-                      imported: importResult.ascents.imported,
-                      skipped: importResult.ascents.skipped,
-                      failed: importResult.ascents.failed,
-                    })}
-                  />
-                </ListItem>
-                <ListItem>
-                  <ListItemText
-                    primary={t('aurora.import.results.attempts')}
-                    secondary={t('aurora.import.results.attemptsSummary', {
-                      imported: importResult.attempts.imported,
-                      skipped: importResult.attempts.skipped,
-                      failed: importResult.attempts.failed,
-                    })}
-                  />
-                </ListItem>
-                <ListItem>
-                  <ListItemText
-                    primary={t('aurora.import.results.circuits')}
-                    secondary={t('aurora.import.results.circuitsSummary', {
-                      imported: importResult.circuits.imported,
-                      skipped: importResult.circuits.skipped,
-                      failed: importResult.circuits.failed,
-                    })}
-                  />
-                </ListItem>
-              </List>
-              {importResult.unresolvedClimbs.length > 0 && (
-                <MuiAlert severity="warning" className={styles.unsyncedAlert}>
-                  <AlertTitle>
-                    {t('aurora.import.results.unresolvedTitle', { count: importResult.unresolvedClimbs.length })}
-                  </AlertTitle>
-                  <div className={styles.unresolvedList}>
-                    {importResult.unresolvedClimbs.slice(0, 20).map((name) => (
-                      <Typography key={name} variant="body2">
-                        {name}
-                      </Typography>
-                    ))}
-                    {importResult.unresolvedClimbs.length > 20 && (
-                      <Typography variant="body2" color="text.secondary">
-                        {t('aurora.import.results.unresolvedMore', {
-                          count: importResult.unresolvedClimbs.length - 20,
-                        })}
-                      </Typography>
-                    )}
-                  </div>
-                </MuiAlert>
-              )}
-            </>
-          )}
-
-          {/* Error phase */}
-          {importPhase === 'error' && importError && (
-            <MuiAlert severity="error">
-              <AlertTitle>{t('aurora.import.dialog.errorTitleShort')}</AlertTitle>
-              {importError}
-            </MuiAlert>
-          )}
-        </DialogContent>
-
-        {/* Actions vary by phase */}
-        {importPhase === 'preview' && (
-          <DialogActions>
-            <Button onClick={handleImportDialogClose}>{t('aurora.import.dialog.cancel')}</Button>
-            <Button variant="contained" onClick={handleImportConfirm}>
-              {t('aurora.import.dialog.confirm')}
-            </Button>
-          </DialogActions>
-        )}
-        {(importPhase === 'complete' || importPhase === 'error') && (
-          <DialogActions>
-            <Button variant="contained" onClick={handleImportDialogClose}>
-              {t('aurora.import.dialog.close')}
-            </Button>
-          </DialogActions>
-        )}
-      </Dialog>
-
-      {/* MoonBoard CSV Import Dialog */}
-      <Dialog
-        open={isMoonBoardImportDialogOpen}
-        onClose={handleMoonBoardImportDialogClose}
-        maxWidth="sm"
-        fullWidth
-        disableEscapeKeyDown={isMoonBoardImporting}
-      >
-        <DialogTitle>{getMoonBoardImportDialogTitle()}</DialogTitle>
-        <DialogContent>
-          {moonBoardImportPhase === 'preview' && moonBoardImportPreview && (
-            <>
-              <Typography variant="body2" color="text.secondary" className={styles.modalDescription}>
-                {moonBoardImportPreview.username
-                  ? t('aurora.moonboard.importDialog.previewIntroWithUser', {
-                      username: moonBoardImportPreview.username,
-                    })
-                  : t('aurora.moonboard.importDialog.previewIntro')}
-              </Typography>
-              <List dense>
-                <ListItem>
-                  <ListItemText
-                    primary={t('aurora.moonboard.importDialog.rows', { count: moonBoardImportPreview.rows })}
-                  />
-                </ListItem>
-                <ListItem>
-                  <ListItemText
-                    primary={t('aurora.moonboard.importDialog.sends', { count: moonBoardImportPreview.sends })}
-                  />
-                </ListItem>
-                <ListItem>
-                  <ListItemText
-                    primary={t('aurora.moonboard.importDialog.flashes', { count: moonBoardImportPreview.flashes })}
-                  />
-                </ListItem>
-                <ListItem>
-                  <ListItemText
-                    primary={t('aurora.moonboard.importDialog.attempts', {
-                      count: moonBoardImportPreview.attempts,
-                    })}
-                  />
-                </ListItem>
-                {moonBoardImportPreview.projects > 0 && (
-                  <ListItem>
-                    <ListItemText
-                      primary={t('aurora.moonboard.importDialog.projects', {
-                        count: moonBoardImportPreview.projects,
-                      })}
-                    />
-                  </ListItem>
-                )}
-                {moonBoardImportPreview.fails > 0 && (
-                  <ListItem>
-                    <ListItemText
-                      primary={t('aurora.moonboard.importDialog.fails', { count: moonBoardImportPreview.fails })}
-                    />
-                  </ListItem>
-                )}
-              </List>
-              <MuiAlert severity="info" className={styles.unsyncedAlert}>
-                {t('aurora.moonboard.importDialog.angleNote', { angle: moonBoardImportPreview.angle })}
-              </MuiAlert>
-              <Typography variant="body2" color="text.secondary">
-                {t('aurora.moonboard.importDialog.previewNote')}
-              </Typography>
-            </>
-          )}
-
-          {moonBoardImportPhase === 'importing' && (
-            <Stack spacing={2} alignItems="center" sx={{ py: 2 }}>
-              <CircularProgress />
-              <Typography variant="body2" color="text.secondary">
-                {moonBoardProgressLabel}
-                {moonBoardImportProgress?.current != null && moonBoardImportProgress.total != null
-                  ? ` (${moonBoardImportProgress.current} / ${moonBoardImportProgress.total})`
-                  : ''}
-              </Typography>
-              {moonBoardImportProgress?.current != null && moonBoardImportProgress.total != null ? (
-                <LinearProgress
-                  variant="determinate"
-                  value={(moonBoardImportProgress.current / moonBoardImportProgress.total) * 100}
-                  sx={{ width: '100%' }}
-                />
-              ) : (
-                <LinearProgress variant="indeterminate" sx={{ width: '100%' }} />
-              )}
-            </Stack>
-          )}
-
-          {moonBoardImportPhase === 'complete' && moonBoardImportResult && (
-            <>
-              {moonBoardImportResult.partialError && (
-                <MuiAlert severity="warning" className={styles.unsyncedAlert}>
-                  <AlertTitle>{t('aurora.moonboard.importResults.partialTitle')}</AlertTitle>
-                  {t('aurora.moonboard.importResults.partialBody')}
-                </MuiAlert>
-              )}
-              <List dense>
-                <ListItem>
-                  <ListItemText
-                    primary={t('aurora.moonboard.importResults.ascents')}
-                    secondary={t('aurora.moonboard.importResults.ascentsSummary', moonBoardImportResult.ascents)}
-                  />
-                </ListItem>
-                <ListItem>
-                  <ListItemText
-                    primary={t('aurora.moonboard.importResults.attempts')}
-                    secondary={t('aurora.moonboard.importResults.attemptsSummary', moonBoardImportResult.attempts)}
-                  />
-                </ListItem>
-              </List>
-              {moonBoardUnresolvedNames.length > 0 && (
-                <MuiAlert severity="warning" className={styles.unsyncedAlert}>
-                  <AlertTitle>
-                    {t('aurora.moonboard.importResults.unresolvedTitle', { count: moonBoardUnresolvedNames.length })}
-                  </AlertTitle>
-                  <div className={styles.unresolvedList}>
-                    {moonBoardUnresolvedNames.slice(0, 20).map((name) => (
-                      <Typography key={name} variant="body2">
-                        {name}
-                      </Typography>
-                    ))}
-                    {moonBoardUnresolvedNames.length > 20 && (
-                      <Typography variant="body2" color="text.secondary">
-                        {t('aurora.import.results.unresolvedMore', {
-                          count: moonBoardUnresolvedNames.length - 20,
-                        })}
-                      </Typography>
-                    )}
-                  </div>
-                </MuiAlert>
-              )}
-            </>
-          )}
-
-          {moonBoardImportPhase === 'error' && moonBoardImportError && (
-            <MuiAlert severity="error">
-              <AlertTitle>{t('aurora.moonboard.importDialog.errorTitle')}</AlertTitle>
-              {moonBoardImportError}
-            </MuiAlert>
-          )}
-        </DialogContent>
-
-        {moonBoardImportPhase === 'preview' && (
-          <DialogActions>
-            <Button onClick={handleMoonBoardImportDialogClose}>{t('aurora.import.dialog.cancel')}</Button>
-            <Button variant="contained" onClick={handleMoonBoardImportConfirm}>
-              {t('aurora.import.dialog.confirm')}
-            </Button>
-          </DialogActions>
-        )}
-        {(moonBoardImportPhase === 'complete' || moonBoardImportPhase === 'error') && (
-          <DialogActions>
-            <Button variant="contained" onClick={handleMoonBoardImportDialogClose}>
-              {t('aurora.import.dialog.close')}
-            </Button>
-          </DialogActions>
-        )}
-      </Dialog>
-    </>
-  );
-}
